Remove CheckoutSession class and embed session model in CheckoutController.
