+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/playground/checkout/CheckoutControllerExampleViewModel.kt
@@ -12,7 +12,7 @@ import androidx.lifecycle.viewModelScope
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
 import com.stripe.android.checkout.CheckoutController
-import com.stripe.android.checkout.CheckoutSession
+import com.stripe.android.checkout.CheckoutController.Session
 import com.stripe.android.checkout.GooglePayConfiguration
 import com.stripe.android.checkout.GooglePayConfiguration.Environment
 import com.stripe.android.paymentelement.CheckoutSessionPreview
@@ -50,7 +50,7 @@ internal class CheckoutControllerExampleViewModel(
         viewModelScope.launch {
             controller.checkoutSession.collect { session ->
                 updateConfiguredState { it.copy(checkoutSession = session) }
-                if (session?.status == CheckoutSession.Status.Complete) {
+                if (session?.status == Session.Status.Complete) {
                     _sessionComplete.tryEmit(Unit)
                 }
             }
@@ -102,7 +102,7 @@ internal class CheckoutControllerExampleViewModel(
     sealed interface Status {
         data object Loading : Status
         data class Configured(
-            val checkoutSession: CheckoutSession?,
+            val checkoutSession: Session?,
         ) : Status
         data class Error(val message: String) : Status
     }
