+++ paymentsheet/src/test/java/com/stripe/android/checkout/CurrencySelectorViewModelTest.kt
@@ -6,6 +6,7 @@ import androidx.test.core.app.ApplicationProvider
 import app.cash.turbine.Turbine
 import app.cash.turbine.test
 import com.google.common.truth.Truth.assertThat
+import com.stripe.android.checkout.CheckoutController.Session
 import com.stripe.android.core.strings.resolvableString
 import com.stripe.android.networking.PaymentAnalyticsRequestFactory
 import com.stripe.android.paymentelement.CheckoutSessionPreview
@@ -162,7 +163,7 @@ internal class CurrencySelectorViewModelTest {
 
     private class Scenario(
         val fakeAnalyticsRequestExecutor: FakeAnalyticsRequestExecutor,
-        val checkoutSessionFlow: MutableStateFlow<CheckoutSession>,
+        val checkoutSessionFlow: MutableStateFlow<Session>,
         val updateCurrencyCalls: Turbine<String>,
         val updateCurrencyResult: Turbine<Result<Unit>>,
         val viewModel: CurrencySelectorViewModel,
